Update the PR a bit to make it "complete"
